Add the possibility of associating ELB target groups to ASG.
When an ASG is totally for one same usage, for example a web app, it is quite common to have an ELB bound with it. But if the ELB was set after the ASG, the target groups can only contain instances that we have to register one by one on creating. If the ELB is registered with the ASG, we can save the registering and let AWS handle that.

Expose the variable lb_target_group_arns mapping to target_group_arns of aws_autoscaling_group.