Bugfix: Empty user_data causing error.
In the case user_data is not used, the default empty string is used. But user_data must be at least 1 char long. So this causes error. And also, if set the user_data to, like " ", it would be treated as compressed data and causing error in cloud-init.log. So give it a script shebang.